arm7tdmi: truncate pipeline[0] to u16 before THUMB_LUT index
record_new_block was indexing the 1024-entry THUMB_LUT with
`(pipeline[0] >> 6) as usize` where pipeline[0] is u32. when stale
upper bits from a prior arm-mode opcode persist (the pipeline state
isnt zeroed across mode flips), the shifted index goes way past 1023
and the bounds check panics.

masked by default codegen ordering so this never fired in shipped
builds. exposed by lto: w/ thin or full lto on, crash within first
~50 frames of pe replay, "len is 1024 but the index is 59025344".

fix is a one-line truncation: read insn as u16 first, then index.
the existing insn-as-u16 cast at the handler dispatch a few lines
later was already there for the same reason — i just hoisted it
earlier in the chain. behaviorally identical for all non-lto builds;
fixes the crash for lto builds.

also adds two release profiles to workspace Cargo.toml for perf
investigation:
  release-cgu1     codegen-units=1 (intra-crate inlining)
  release-thinlto  codegen-units=1 + lto="thin" (cross-crate)
both inherit from release. cgu1 didnt show measurable gain over
default release on cached_interp-only build (PE 465 vs 472 median);
thinlto still panics on a separate pre-existing bug not addressed
here (cpu in thumb mode at bios addr 0x16 — needs deeper audit).

+++ arm7tdmi/src/cpu.rs
@@ -1716,7 +1716,12 @@ impl<I: MemoryInterface> Arm7tdmiCore<I> {
             if entry_thumb {
                 let pc = self.pc & !1;
                 let fetched_now = self.load_16(pc, self.next_fetch_access);
-                let insn = self.pipeline[0];
+                // pipeline[0] is u32; in thumb mode only the low 16 bits
+                // are the live opcode. truncate before LUT index — without
+                // this, stale upper bits from a prior arm-mode block can
+                // index the 1024-entry LUT well past its end. usually
+                // hidden by codegen ordering; exposed by LTO.
+                let insn = self.pipeline[0] as u16;
                 self.pipeline[0] = self.pipeline[1];
                 self.pipeline[1] = fetched_now as u32;
                 let handler = Self::THUMB_LUT[(insn >> 6) as usize].handler_fn;
